+++ spec/requests/stash_api/urls_controller_spec.rb
@@ -1,8 +1,6 @@
 require 'rails_helper'
 require 'uri'
 require_relative 'helpers'
-require 'fixtures/stash_api/metadata'
-require 'fixtures/stash_api/curation_metadata'
 require 'cgi'
 require 'digest'
 # see https://relishapp.com/rspec/rspec-rails/v/3-8/docs/request-specs/request-spec
