+++ spec/requests/stash_api/versions_controller_spec.rb
@@ -1,7 +1,5 @@
 require 'uri'
 require_relative 'helpers'
-require 'fixtures/stash_api/metadata'
-require 'fixtures/stash_api/curation_metadata'
 require 'cgi'
 
 # see https://relishapp.com/rspec/rspec-rails/v/3-8/docs/request-specs/request-spec
